refactor: Restructure project directory
