+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".ui.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -23,8 +23,8 @@
             android:exported="false" />
 
         <activity
-            android:name=".EditorActivity"
-            android:parentActivityName=".MainActivity" />
+            android:name=".ui.EditorActivity"
+            android:parentActivityName=".ui.MainActivity" />
     </application>
 
 </manifest>
